requirements_quick.txt
# AgriBot Pro - Quick Start Requirements
# Install with: pip install -r requirements_quick.txt

# Core Framework
streamlit>=1.30.0
streamlit-option-menu>=0.3.6

# Data & Visualization
pandas>=2.0.0
plotly>=5.18.0

# API & Web
requests>=2.31.0

# Database
sqlalchemy>=2.0.0

# LLM and AI
langchain>=0.2.0
langchain-ollama>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0

# Vector Store
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Document Processing
pypdf>=3.17.0